chore: move alias to its own file
